requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.6
marshmallow==3.26.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.96.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
propcache==0.3.2
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.15
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
